// angular/frontend/src/app/models/auth-login.model.ts
export interface AuthLoginModel {
    username: string;
    password: string;
  }

export interface AuthLoginResponseModel {
    refresh: string;
    access: string;
};

export interface AuthLoginResponseFailureModel {
  detail: string;
};

export interface AuthTokenRefreshResponseModel {
  access: string;
};